modified components with remove button
