+++ packages/client/src/components/ViewCartProducts/ViewCartProducts.js
@@ -5,7 +5,7 @@ import PropTypes from 'prop-types';
 export const ViewCartProducts = ({ label }) => {
   return (
     <div className="Cart-products">
-      <span>Cart(3)</span>
+      <span>Cart (3)</span>
     </div>
   );
 };
